# routes/api-docs/apiswagger.yaml
swagger: '2.0'
info:
  description: Triumf API service
  version: 1.0.0
  title: Triumf API
  contact:
    email: prokopenko.cnw@gmail.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
host: http://193.200.173.204/
basePath: 
paths:
...
  api/v4.2/ETH/sendRawTransaction/{rawTransaction}:
    get:
      tags:
      - sendRawTransaction
      summary: Send and broadcast raw transaction
      operationId: sendRawTransaction
      consumes:
      - application/json
      produces:
      - application/json
      parameters:
      - name: rawTransaction
        in: path
        description: The serialized raw transaction in hex format
        required: true
        type: string
      responses:
        200:
          description: Success execute
          schema:
            $ref: '#/definitions/hash'
        500:
          description: Service unavialable
  api/v4.2/ETH/getTransactionsList/{address}:
    get:
      tags:
      - getTransactionsList
      summary: Get address transactions
      operationId: getTransactionsList
      consumes:
      - application/json
      produces:
      - application/json
      parameters:
      - name: address
        in: path
        description: The address for ETH wallet
        required: true
        type: string
      responses:
        200:
          description: Success execute
          schema:
            $ref: '#/definitions/transactionsList'
        500:
          description: Service unavialable
...
  api/v4.2/ETH/getBalance/{address}:
    get:
      tags:
      - getBalance
      summary: Get address balance
      operationId: getBalance
      consumes:
      - application/json
      produces:
      - application/json
      parameters:
      - name: address
        in: path
        description: The address for ETH wallet
        required: true
        type: string
      responses:
        200:
          description: Success execute
          schema:
            $ref: '#/definitions/balance'
        500:
          description: Service unavialable
...
definitions:
  transactionsList:
    type: object
    description: ETH transactions object, in - input transactions array, out - output transactions array, pending_in - pending input transactions array, pending_out - pending outcome transactions array
    properties:
      in:
        type: array
        items:
          $ref: '#/definitions/transactionETH'
      out:
        type: array
        items:
          $ref: '#/definitions/transactionETH'
      pending_in:
        type: array
        items:
          $ref: '#/definitions/transactionETH'
      pending_out:
        type: array
        items:
          $ref: '#/definitions/transactionETH'    
  transactionETH:
    type: object
    properties:
        hash:
          type: string
        timestamp:
          type: string
        from:
          type: string
        to:
          type: string
        value:
          type: string
        fee:
          type: string
        blockNum:
          type: integer
        input:
          type: object
          properties:
            to:
              type: string
            value:
              type: string
            from:
              type: string
        status:
          type: boolean
  hash:
    type: object
    properties:
      hash:
        type: string
        description: the transaction hash, or the zero hash if the transaction is not yet available
...
  balance:
    type: object
    properties:
      balance:
        type: string
        description: balance value in wei, string of int format
...