# routes/api/v4.2/btc/index.js
const express = require('express');
const app = module.exports = express();
const btcController = require(`../../../../controllers/btcController`);
/**
  * @swagger
  * definitions:
  *   Transactions:
  *     required:
  *       - in
  *       - out
  *     properties:
  *       in:
  *         type: object
  *       out:
  *         type: object
*/

/**
 * @swagger
 * /getTransactionsList/address:
 *   get:
 *     description: Transactions list
 *     produces:
 *       - application/json
 *     parameters:
 *       - name: address
 *         description: Address ETH to get Transaction List.
 *         in: formData
 *         required: true
 *         type: string
 *     responses:
 *       200:
 *          description: Transactions List
 *          schema:
 *              type: object
 *              $ref: '#/definitions/Transactions'
 *         
 *         
 */

// send raw transaction
app.get('/sendRawTransaction/:raw', (req, res, next) =>{
    const raw = req.params.raw;
    btcController.sendRawTransaction(raw)
        .then(response => {
            res.send(response);
        })
        .catch(error => {
            next(error);
        })
});
// get balance by address
app.get('/getBalance/:address', (req, res, next) => {
    const address = req.params.address;
    btcController.getBalanceNew(address)
        .then(balance => {
            res.send(balance);

        })
        .catch(error => {
            next(error)
        })
});
// get balanceOld by address
app.get('/getBalanceOld/:address', (req, res, next) => {
    const address = req.params.address;
    btcController.getBalance(address)
        .then(balance => {
            res.send(balance);

        })
        .catch(error => {
            next(error)
        })
});
// get listUTXOs by address
app.get('/getUTXOs/:address', (req, res, next) => {
    const address = req.params.address;
    btcController.getUTXOsP(address, 0)
        .then(response => {
            res.setHeader('UtxoPages', response.pages);
            res.send(response);
        })
        .catch(error => {
            next(error)
        })
});
// get listUTXOs by address & pages
app.get('/getUTXOs/:address/:page', (req, res, next) => {
    const page = parseInt(req.params.page);
    const address = req.params.address;
    btcController.getUTXOsP(address, page)
        .then(response => {
            res.send({
                utxos: response.utxos
            });
        })
        .catch(error => {
            next(error)
        })
});
//get txs list by address
app.get('/getTransactionsList/:address', (req,res, next) => {
    const address = req.params.address;
    btcController.getAllTxList(address, 0)
        .then(transactions => {
            res.setHeader('TrPages', transactions.pages);
            res.send(transactions);
        })
        .catch(error => {
            next(error)
        })
});
//get txs list by address by page
app.get('/getTransactionsList/:address/:page', (req,res, next) => {
    const page = parseInt(req.params.page);
    const address = req.params.address;
    btcController.getAllTxList(address, page)
        .then(transactions => {
            res.send({
                pending:        transactions.pending,
                transactions:   transactions.transactions
            });
        })
        .catch(error => {
            next(error)
        })
});
// get raw transaction by id
app.get('/getTransactionById/:txid', (req, res, next) => {
    const txid = req.params.txid;
    btcController.getTransactionById(txid)
        .then(tx => {
            res.send(tx);
        })
        .catch(error => {
            next(error)
        })
});